fix: make checkboxes action items
In the current state of this issue template, it's hard for me to keep track of what actually needs to be done to complete the burial of the item. What about trying to make it easier on the gravediggers by making every checkbox an action item that needs to be completed before the issue is close?